pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lucidtree"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "fonttools==4.61.1",
    "iniconfig==2.3.0",
    "isort>=7.0.0",
    "kiwisolver==1.4.9",
    "librt==0.7.5",
    "matplotlib==3.10.8",
    "mypy==1.19.1",
    "mypy-extensions==1.1.0",
    "numpy==2.3.5",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pillow==12.1.1",
    "pip-audit>=2.10.0",
    "pluggy==1.6.0",
    "pygame>=2.6.1",
    "pygments==2.19.2",
    "pyparsing==3.2.5",
    "pytest==9.0.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "ruff==0.14.10",
    "sgfmill==1.1.1",
    "six==1.17.0",
    "termcolor>=3.3.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "typing-extensions==4.15.0",
    "urllib3==2.6.3",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest]
minversion = "9.0"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
files = ["src/**/*.py", "api/**/*.py"]
strict = true
ignore_missing_imports = true
mypy_path = ["src", "api"]
